Added sessions on server

+++ web-applications-dotnet/Controllers/CustomerController.cs
@@ -1,4 +1,5 @@
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using web_applications_dotnet.DAL;
@@ -13,6 +14,8 @@ namespace web_applications_dotnet.Controllers
 
         private readonly ILogger<CustomerController> _log;
 
+        private const string LoggedIn = "logedIn";
+
         public CustomerController(ICustomerRepository db, ILogger<CustomerController> log)
         {
             _db = db;
@@ -21,6 +24,10 @@ namespace web_applications_dotnet.Controllers
         
         public async Task<ActionResult> Save(Customer customer)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(LoggedIn)))
+            {
+                return Unauthorized();
+            }
             if (ModelState.IsValid)
             {
                 var ret = await _db.Save(customer);
@@ -37,12 +44,20 @@ namespace web_applications_dotnet.Controllers
 
         public async Task<ActionResult> GetAll()
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(LoggedIn)))
+            {
+                return Unauthorized();
+            }
             var list = await _db.GetAll();
             return Ok(list);
         }
 
         public async Task<ActionResult> Delete(int id)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(LoggedIn)))
+            {
+                return Unauthorized();
+            }
             var ret = await _db.Delete(id);
 
             if (ret) return Ok("Customer deleted");
@@ -53,6 +68,10 @@ namespace web_applications_dotnet.Controllers
 
         public async Task<ActionResult> GetOne(int id)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(LoggedIn)))
+            {
+                return Unauthorized();
+            }
             var customer = await _db.GetOne(id);
 
             if (customer != null) return Ok(customer);
@@ -63,6 +82,10 @@ namespace web_applications_dotnet.Controllers
 
         public async Task<ActionResult> Update(Customer customer)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(LoggedIn)))
+            {
+                return Unauthorized();
+            }
             if (ModelState.IsValid)
             {
                 var ret = await _db.Update(customer);
@@ -81,12 +104,22 @@ namespace web_applications_dotnet.Controllers
             if (ModelState.IsValid)
             {
                 var ret = await _db.LogIn(user);
-                if (ret) return Ok(true);
+                if (ret)
+                {
+                    HttpContext.Session.SetString(LoggedIn, "LoggedIn");
+                    return Ok(true);
+                }
                 _log.LogInformation("Log in failed for user: "+user.Username);
+                HttpContext.Session.SetString(LoggedIn, "");
                 return Ok(false);
             }
             _log.LogInformation("Input validation failed");
             return BadRequest("Input validation failed on server");
+        }
+
+        public void LogOut()
+        {
+            HttpContext.Session.SetString(LoggedIn, "");
         }
     }
 }
